Decouple score calculation from database
